added save/load between scenes kinda and other misc
Added the ability to respawn where your last save point was kinda, it's
buggy with mutliple instances of GM forming, and puts you at the
original position that the character was placed, but works if you
respawn again.  Also fixed animations and added the ability to switch
between scenes, but only one way for now for testing.  Finally ,added a
death screen