Respawn bug fixed
everything up until this point should be working, although some of it
might be a bit "hacky".  I will try and optimize stuff for the least
amount of work later on, namely the new game / load game on main menu
becuase I kind of rushed that.  Thanks Sym for finding my rartness